+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Header/Footer widget added.
+* Optimize builder with index
+* Empty loader lines size added for both grid and list view
+
 ## 0.0.1
 
 * Initial version.